take database form real provider

+++ src/Models/Location.kt
@@ -1,3 +1,3 @@
 package com.amerharb.shape.models
 
-data class Location(val location: String, val tempUnit: TemperatureUnit, val nextDaysTemp: List<Int>)
+data class Location(val location: String, val tempUnit: TemperatureUnit, val nextDaysTemp: List<Float>)
